Suppress Qt deprecation error
Passing a `QMutex::RecursionMode` parameter to the `QMutex` constructor
is superseded by `QRecursiveMutex`, which was introduced in Qt 5.14.
We need to support Qt 5.12, which was released in December 2018 and is
included in Ubuntu 20.04.

As a workaround, suppress the error and make explicit the minimum Qt
version required to build LogWitch.