Use the repository-provided `version.cmake` as fallback
This fixes the build when using the source tarball. Since tarballs
downloaded from GitHub do not contain Git version information,
`getgitversion.cmake` is not capable of generating a valid
`version.cmake` from the Git tag.

This change implies that `version.cmake` must be kept up to date when
tagging a new release.