Fix undefined behaviour in destruction of `QuickSearchBar`
Reported by UndefinedBehaviorSanitizer.

Since all `QuickSearchBar` objects are owned by `LogEntryTableWindow`,
they are destroyed after it. Hence, calling a method of `m_logWindow` in
the destructor of `QuickSearchBar` yields undefined behaviour.

This is not a use-after-free bug because the memory of the
`LogEntryTableWindow` object has not been deallocated just yet - but its
destructor has already been called.

To fix this bug, hold a `std::shared_ptr` to the `RuleTable` as a member
variable in `QuickSearchBar`. This eliminates the need to call
`m_logWindow->getRuleTable()`.