Fix segmentation fault when closing a log window, then opening another
This bug was introduced by commit
e261d7bd1fc01498f5eaa44a347b455531acab4e, which changed a `static_cast`
to a `dynamic_cast`.

As it turns out, one cannot `dynamic_cast` an object that is in the
process of being destroyed - with control moving up the chain of virtual
destructors - to a subclass whose destructor has already been invoked.
As a result, `view` equals `nullptr` and the child object is not purged
from the group.

To fix this bug, remove the harmful `dynamic_cast`; `std::find()` works
without a cast.