Fix undefined behaviour in destruction of `LogfileAnalyser`
Reported by UndefinedBehaviorSanitizer.

The `LogEntryTableWindow` objects are (transitively) owned by
`LogfileAnalyser`, so there are two possible situations when a
`LogEntryTableWindow` gets destroyed:
1. The user closes the window -> `LogfileAnalyser` is still alive.
2. `LogfileAnalyser` is being shut down -> it is no longer alive.

The first case is covered by the `QObject::destroyed()` signal.

The second case has not been covered so far. It results in
`subWindowDestroyed()` being invoked on a half-destroyed
`LogEntryTableWindow` object. This is not a use-after-free bug but still
undefined behaviour.

Fix this bug by keeping a list of signal-to-slot connections and
disconnecting them in the destructor.